rename+add requirements.txt
